Fix PR 226 CI dependency and module gates
